Fixed suspender issue: Keep track of effective_bol (beginning of line):
Remember if previously removed token was at beginning of line. This will help removing header_atx tokens:

Before, this header hash mark would not have been removed: `@# header` because the token is required to be at beginning of line.

Now, the subtitle will be removed first, pushing the hash mark to the beginning of line, and the hash mark will be removed as well.