Updated Suspension::Token and Suspender:
There was a but where inline IALs were detected as block IALs resulting in inconsistent behavior around newlines after headers: If a header had an inline IAL like "# header{: .italic}\n\n" that inline IAL was detected as a block IAL and the trailing newline was consumed. However a header without an inline IAL at the end of the line did not have the trailing newline removed. So we got inconsistent behavior every time we used suspension.

Now inline IALs are detected correctly and the behavior is consistent. We accomplish this by making only subtitle_marks and gap_marks transparent to beginning of line. We need it so that atx header preceded by a subtitle_mark are detected as such, even though the hash sign is not really at the beginning of the line.