Fixed issues where utf8 multibyte chars with identical start, end, or middle bytes confused the dmp algorithm, producing invalid utf8 byte sequences:
Invalid UTF-8 byte sequences in diff can be caused, e.g., by two different multibyte characters at the same position where the first bytes are identical, and a subsequent one is different. DMP splits that multibyte char into separate diff units and thus creates invalid UTF8 byte sequences.

Example: "a—b" and "a…b"

[
  [0, "a\xE2\x80"],
  [-1, "\x94b"],
  [1, "\xA6b"]
]

This is normally not a problem since combining the diff units will result in valid utf8 byte sequences. However, in suspension we need to measure diff string lengths for token suspension and unsuspension, so the broken utf8 byte sequences get in the way and we need to fix them.

We fix these invalid diffs by moving separated byte sequences back together.

We already addressed this issue previously in StringComparer. This commit improves on the original fix and moves it to the source of the problem (DiffAlgorithm) so that all code that relies on dmp will now get valid (utf8 aware) diff strings.